Added functionality to exportv2 to separate out files that require extra attention

+++ download/resource_write.go
@@ -1,7 +1,9 @@
 package download
 
 import (
+	"errors"
 	"os"
+	"strings"
 
 	"github.com/dtcookie/hcl"
 	"github.com/dynatrace-oss/terraform-provider-dynatrace/hclgen"
@@ -10,6 +12,10 @@ import (
 func (me ResourceData) WriteResourceSeparate(dlConfig DownloadConfig, resName string, resFolder string, resources Resources, resNameCnt NameCounter) error {
 	var err error
 	for _, resource := range resources {
+		if resource.ReqInter {
+			continue
+		}
+
 		var file *os.File
 		fileName := dlConfig.TargetFolder + "/" + resFolder + "/" + resFolder + "." + escf(resource.Name) + ".tf"
 		os.Remove(fileName)
@@ -34,7 +40,6 @@ func (me ResourceData) WriteResourceSeparate(dlConfig DownloadConfig, resName st
 				file.Close()
 				return err
 			}
-
 		}
 		file.Close()
 	}
@@ -59,6 +64,49 @@ func (me ResourceData) writeDashboardSharing(file *os.File, name string) error {
 	if err := hclgen.Export(restObject, file, "dynatrace_dashboard_sharing", escape(name)); err != nil {
 		file.Close()
 		return err
+	}
+	return nil
+}
+
+func (me ResourceData) WriteResReqAttn(dlConfig DownloadConfig) error {
+	var err error
+	for resName := range InterventionInfoMap {
+		if resources, exists := me[resName]; exists {
+			for _, resource := range resources {
+				if !resource.ReqInter {
+					continue
+				}
+
+				folderName := dlConfig.TargetFolder + "/" + ".requires_attention"
+				if _, err := os.Stat(folderName); errors.Is(err, os.ErrNotExist) {
+					err := os.Mkdir(folderName, os.ModePerm)
+					if err != nil {
+						return err
+					}
+				}
+
+				var file *os.File
+				fileName := folderName + "/" + strings.TrimPrefix(resName, "dynatrace_") + "." + escf(resource.Name) + ".tf"
+				os.Remove(fileName)
+				if file, err = os.Create(fileName); err != nil {
+					return err
+				}
+
+				if dlConfig.CommentedID {
+					if err := hclgen.Export(resource.RESTObject, file, resName, escape(resource.Name), "id = "+resource.ID); err != nil {
+						file.Close()
+						return err
+					}
+				} else {
+					if err := hclgen.Export(resource.RESTObject, file, resName, escape(resource.Name)); err != nil {
+						file.Close()
+						return err
+					}
+				}
+
+				file.Close()
+			}
+		}
 	}
 	return nil
 }
